[Update] Optimize the generation of DFG.
Change the logic of DFG generation. Now in DFG, red line stands for non-inter-basicblock and non-inter-iteration data flow, and green line stands for inter-basicblock or inter-iteration flow. DFGNodeName in DFG now is DFGNodeName_number, and number is the ID of the basicblock that DFGNode belongs to. Also changed the generation sequence of DFGNode. Now the ID of DFGNode is the same as the order of LLVM IR.